TypeScript Builder refactor: extract PipeChainEmitter
Move the pipe (|>) lowering logic out of TypeScriptBuilder into a
dedicated helper at lib/backends/typescriptBuilder/pipeChainEmitter.ts.

Behaviour change: none. The emitter holds the pipe-temp counter and
calls back into the builder for the bits it cannot do itself
(processNode, buildAssignmentLhs, buildStateConfig,
generateFunctionCallExpression, str, scopes). At the call sites the
builder now reads declaratively:

  this.pipes.tryGetChainStages(stmt)
  this.pipes.expand(stmt, stages, baseId)
  this.pipes.bind(leftIR, stage)

Methods moved (all previously on the builder):
  - _pipeCounter
  - getPipeChainStages   -> PipeChainEmitter.tryGetChainStages
  - buildPipeBind        -> PipeChainEmitter.bind
  - expandPipeChain      -> PipeChainEmitter.expand
  - buildPipeLambda      -> private
  - buildPipeStageBody   -> private
  - processValueAccessPartial -> private (only the pipe code used it)

Builder shrinks from 3658 to 3421 lines.

Verification: - pnpm exec tsc --noEmit (clean)
  - pnpm run build + make stdlib
  - pnpm test:run (3865/3865)
  - All pipe fixture tests (find tests -path "*pipe*" -name "*.test.json")
  - pnpm run lint:fix + pnpm run lint:structure (clean)
Amp-Thread-ID: https://ampcode.com/threads/T-019e3b3d-f41a-7078-94bb-e29ff6740981
